Initial commit: a django web app with a failing test
Creates a blank django web app with a single app
called "window" that will be used to query and display
the results of searches of books.

The failing test is a functional test expecting there
to be an input element on the homepage.

The steps for working on this project and running the
tests are in the README.