Display an input element on the homepage
Passes the first function test expecting there to be
a search input on the homepage.

Renders the "window/templates/window/index.html" template.

Includes the "window" app in settings.py so that the
template discovery works properly.

Fixes a bug in the selenium test, using the ".get_attribute"
method on web elements.