First commit of a new branch called vagrant-vm
I ran `vagrant up` to create a default Vagrantfile,
and filled it in with my settings.

Ubuntu 16.04 has python 3.5 installed. This commit
downgrades the project version of python to 3.5.
The .python-version file is read by pyenv.
The Pipfile is read by pipenv.

The Django app is "deployed" by syncing the local "bookstorewindow"
directory on the VM.

I added an Ansible playbook "configure.yml" for
configuring the VM after a fresh install.

I ran into issues with running Pipenv commands
as an Ansible task. I think a better way is to
use the Ansible pip module with the virtualenv
params.