Use unbound models to hold book data
I'm using Django's models to hold book data even though
I won't be persisting the data to a db.

I implemented a custom model manager that will be used
to query the Google Books API and return the results as
Book objects.

A queryset of Books is now passed to the template so the
results can be rendered on the page.

The first user test now passes, expecting a user to search
for "The Bible" and return more than 0 results.