Add authentication to Google Developers Console
I obtained an API key to the Google Developers Console,
and stored it in a yaml file. I encrypted that file with
the ansible-vault tool.

The reason I used ansible-vault for encryption was to make it
easy to eventually deploy the API key on servers using ansible.

I also added a script that writes an ".env" file. Variables defined
in ".env" are loaded by pipenv every time the venv is activated.
Running bin/setup-env.py prompts the user for the required environment
variables.

The ".env" file is not tracked.