Return results to queries from the Google Books API
This commit is the first time the app has felt like
it was working.

Connecting to the Google Books API requires an API key.
In order to work, the application expects an environment variable
named GOOGLE_API_KEY. This commit populates that environment
variable by reading the value from the ansible-vault and
writing it to the ".env" file written by the bin/setup-env.py
script.

Testing the results of calls to the Google Books API
should not require lots of repeat requests. This commit
introduces the betamax package which integrates nicely
with the requests library to retrieve real responses once,
and store these in a fixture for repeat requests.

The cassettes should not be committed to version control
because they contain the application's API key in plain
text.