Use an Ansible playbook to create a Digital Ocean droplet
The provision.yml playbook can be used to create a droplet for
staging the app, and eventually for production.

Using the digital_ocean ansible module requires "dopy". However,
the version of dopy on PyPI is not compatible with python3.
As a workaround, I installed a fork of dopy that has been
updated to work with python3.

$ pip install git+https://github.com/rolylolypoly/dopy