Refactor window app
This commit refactors the code that touches the Google Books API
to more cleanly separate the models used in this app from the
external API. This allows for cleaner testing of the model and
app behavior that is separate from parsing the results received
from the Google Books API.

I also fixed up the tests to correctly mock out any live requests
sent to the Google Books API. Now the tests correctly use Betamax
to cache the results after they have been requested once. This means
that the unit tests can be run without an internet connection.

This commit introduces a new model: Volume. The issue I was having
is that the Google Books API has multiple Volume entities that are
for all intents and purposes identical. I will use the Book model
to stand for what a normal human would consider to be a book, and
a related model called Volume that has a unique google book id.

*This is the first time a DB is being used in this app.* Before
everything was done in memory. Now model objects are being saved
to the DB so that the ForeignKeys between Books and Volumes are
upheld.