Improve README to document previously missing dependencies
The fact that the user tests require a separate webdriver to
be installed is now documented. Developers working on this project
should install the `chromedriver` to run headless user tests.

$ brew cask install chromedriver

The full version of python that is being used is now fully documented
with pyenv and pipenv. pyenv installs 3.7.1, which is the version
of python available on Heroku machines. The pipenv creates a virtualenv
using the expected version of python.

The heroku documentation was updated to document how to connect
to an existing app when a new one isn't being created.

The required environment variables are now documented in the README.