Create an app containing all logic related to Google Books
This commit pulls the code that was in the "window" app that
deals with connecting to the Google Books API and puts it in
a separate app, "google_books".

The main concern here is to decouple the tests that require
a valid Google API Key from those that are just running the
app internals.

Eventually we can use `model_mommy` for testing the window
app without worrying about a Google API Key.