Test the app on a Vagrant VM
This commit introduces Vagrant to the development cycle.
I installed VirtualBox and Vagrant for testing on a VM.
I am using ansible to deploy the app on the VM.

The VM I'm using is a Ubuntu 18.04 box from Jeff Geerling, which has
python 3.6.7 on it. I downgraded the python version to python 3.6.7.

Getting the user tests to run on the VM required installing
the chromium driver and providing the path to the `chromedriver`
executable.